HEAT_create_dimension_table {heatdata}	R Documentation

Creates the dimension table including adding the colors and shapes for dimension subgroups

Description

Creates the dimension table including adding the colors and shapes for dimension subgroups

Usage


HEAT_create_dimension_table(.data, cutoff_for_single_color = 7)

Arguments

cutoff_for_single_color
if the number of subgroups in a dimension is less than or equal to this number then each subgroup will get a different color. Otherwise all subgroups get the same color. So if the cutoff is 7 all dimensions with 7 or fewer subgroups will get different colors (either sequential or qualitative) all dimensions with 8 or more subgroups will be assigned a single color

[Package heatdata version 0.0.1 Index]