HEAT_create_setyrsrc_table {heatdata}	R Documentation

Create a table with the unique set of setting, year and source

Description

the actual names of the setting, year and source variables (which might be setting, year and source) are in the heatmeasures::setting_year_source_variables vector

Usage


HEAT_create_setyrsrc_table(.data)

Arguments

.data

[Package heatdata version 0.0.1 Index]