HEAT_create_subregionminmax_table {heatdata}	R Documentation

Compute the table of subregion min and max values for the mapping (only required for HEAT)

Description

Compute the table of subregion min and max values for the mapping (only required for HEAT)

Usage


HEAT_create_subregionminmax_table(.data)

Arguments

.data

[Package heatdata version 0.0.1 Index]