HEAT_rename_variables {heatdata}	R Documentation

Rename a table's variables

Description

Occasionally the database we've received has a slightly different name than the app expects and we need to alter it to fit.

Usage


HEAT_rename_variables(.data, table_type = "all")

Arguments

table_type

[Package heatdata version 0.0.1 Index]