spatial_fix_survey_years {heatdata}	R Documentation

Change the DHS shapefile year so that the year matches the WHO HEAT year – for DHS surveys that spanned two years, DHS may have assigned one year while WHO HEAT assigned the other year

Description

Change the DHS shapefile year so that the year matches the WHO HEAT year – for DHS surveys that spanned two years, DHS may have assigned one year while WHO HEAT assigned the other year

Usage


spatial_fix_survey_years(.data)

Arguments

.data

[Package heatdata version 0.0.1 Index]